Refactor API services
