+++ models/test-device.ts
@@ -1,0 +1,5 @@
+export type TestDevice = {
+  deviceId: string;
+  deviceType: string;
+  owner: string;
+};
